refactor: split simulation output segments
