+++ projects/policyengine-api-simulation/tests/test_standalone_simulation_contract.py
@@ -15,7 +15,15 @@ PACKAGED_RUNTIME_MODULES = (
     "policyengine_api_simulation.release_bundle",
     "policyengine_api_simulation.simulation",
     "policyengine_api_simulation.simulation_macro_output",
+    "policyengine_api_simulation.simulation_output_budget",
     "policyengine_api_simulation.simulation_output_builder",
+    "policyengine_api_simulation.simulation_output_cliff",
+    "policyengine_api_simulation.simulation_output_common",
+    "policyengine_api_simulation.simulation_output_distribution",
+    "policyengine_api_simulation.simulation_output_geographic",
+    "policyengine_api_simulation.simulation_output_inequality",
+    "policyengine_api_simulation.simulation_output_labor",
+    "policyengine_api_simulation.simulation_output_poverty",
     "policyengine_api_simulation.simulation_runtime",
     "policyengine_api_simulation.telemetry",
 )
